adicion de paquetes

+++ lib/pages/client/reservas_realizadas/reservas_pendientes.dart
@@ -60,7 +60,7 @@ class ReservasPendientesCliente extends StatelessWidget {
               model: data['vehiculo']['marcaVehiculo'],
               plate: data['vehiculo']['placaVehiculo'],
               status: data['estado'],
-              total: data['total'],
+              total: data['total'].toDouble(),
               typeVehicle: data['vehiculo']['tipo'],
               id: document.id,
               idDuenio: data['parqueo']['idDuenio'],
